+++ server/base/Client.h
@@ -787,7 +787,7 @@ private:
 
     uint32_t tryCapture(const uint16_t &item);
     bool changeOfMonsterInFight(const uint8_t &monsterPosition);
-    void confirmEvolutionTo(PlayerMonster * playerMonster,const uint32_t &monster);
+    void confirmEvolutionTo(PlayerMonster * playerMonster,const uint16_t &monster);
     std::vector<uint8_t> addPlayerMonster(const std::vector<PlayerMonster> &playerMonster);
     std::vector<uint8_t> addPlayerMonster(const PlayerMonster &playerMonster);
     bool addPlayerMonsterWithChange(const PlayerMonster &playerMonster);
